+++ plotnine/_mpl/transforms.py
@@ -1,0 +1,3 @@
+from matplotlib.transforms import Bbox
+
+ZEROS_BBOX = Bbox([[0, 0], [0, 0]])