+++ plotnine/options.py
@@ -20,14 +20,9 @@ dpi = 100
 #: Default figure size inches
 figure_size = (640 / dpi, 480 / dpi)
 
-#: Default parameters for how to tune the subplot layout
-# Choosen to match MPL 2.0 defaults
-SUBPLOTS_ADJUST = {
-    "left": 0.125,  # the left side of the subplots of the figure
-    "right": 0.9,  # the right side of the subplots of the figure
-    "bottom": 0.11,  # the bottom of the subplots of the figure
-    "top": 0.88,  # the top of the subplots of the figure
-}
+#: A size that is proportional of the figure width and
+#: is used by some themes to determine other margins
+base_margin = 0.01
 
 
 def get_option(name: str) -> Any: