+++ plotnine/themes/theme_classic.py
@@ -17,16 +17,12 @@ class theme_classic(theme_bw):
     """
 
     def __init__(self, base_size=11, base_family=None):
-        theme_bw.__init__(self, base_size, base_family)
+        super().__init__(base_size, base_family)
         self += theme(
             panel_border=element_blank(),
             axis_line=element_line(color="black"),
-            panel_grid_major=element_line(),
-            panel_grid_major_x=element_blank(),
-            panel_grid_major_y=element_blank(),
-            panel_grid_minor=element_line(),
-            panel_grid_minor_x=element_blank(),
-            panel_grid_minor_y=element_blank(),
+            panel_grid_major=element_blank(),
+            panel_grid_minor=element_blank(),
             strip_background=element_rect(colour="black", fill="None", size=1),
             legend_key=element_blank(),
         )